sp2txtbot/requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Babel==2.9.1
boto3==1.26.27
botocore==1.29.27
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coloredlogs==15.0.1
coverage==6.5.0
cryptography==38.0.4
exceptiongroup==1.0.4
flake8==6.0.0
frozenlist==1.3.3
grpcio==1.51.1
grpcio-tools==1.48.2
humanfriendly==10.0
idna==3.4
iniconfig==1.1.1
iso8601==0.1.16
jmespath==1.0.1
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.3
mypy==0.991
mypy-extensions==0.4.3
packaging==22.0
pluggy==1.0.0
protobuf==3.19.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
pydub==0.25.1
pyflakes==3.0.1
PyJWT==2.6.0
pypika-tortoise==0.1.6
pytest==7.2.0
pytest-asyncio==0.20.3
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
speechkit==2.1.1
tomli==2.0.1
tortoise-orm==0.18.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
